Extract AST submodules
